+++ test/integration/integration/recaptcha_validation_interceptor_test.clj
@@ -1,7 +1,7 @@
 (ns integration.recaptcha-validation-interceptor-test
   (:require [cheshire.core :as json]
             [clj-http.fake :as http.fake]
-            [clojure.test :refer :all]
+            [clojure.test :refer [is testing]]
             [com.stuartsierra.component :as component]
             [common-clj.component.config :as component.config]
             [common-clj.component.helper.core :as component.helper]